use actions for contestlive
