+++ lib/quadquizaminos_web/router.ex
@@ -39,8 +39,8 @@ defmodule QuadquizaminosWeb.Router do
     live "/leaderboard", LeaderboardLive
     live "/leaderboard/:board_id", LeaderboardLive.Show
     live "/contestboard", ContestboardLive
-    live "/contests", ContestsLive
-    live "/contests/:contest_id", ContestsLive.Show
+    live "/contests", ContestsLive, :index
+    live "/contests/:contest_id", ContestsLive, :show
     live "/contest_rules", ContestRules
     live "/contest_prizes", ContestPrizes
     get "/anonymous", PageController, :anonymous
