+++ lib/quadquizaminos_web/templates/layout/navbar.html.leex
@@ -3,7 +3,7 @@
          <li><a href="/">Home</a></li>
          <li><a href= <%= Routes.page_path(@conn, :instructions) %> >How to Play</a></li>
          <li><a href= <%= Routes.live_path(QuadquizaminosWeb.Endpoint, QuadquizaminosWeb.LeaderboardLive) %> >Leader board</a></li>
-         <li><a href= <%= Routes.live_path(QuadquizaminosWeb.Endpoint, QuadquizaminosWeb.ContestsLive) %> >Contests</a></li>
+         <li><a href= <%= Routes.contests_path(QuadquizaminosWeb.Endpoint, :index) %> >Contests</a></li>
          <li><a href= <%= Routes.live_path(QuadquizaminosWeb.Endpoint, QuadquizaminosWeb.ContestPrizes) %> >Contest prizes</a></li>
          <li><a href= <%= Routes.live_path(QuadquizaminosWeb.Endpoint, QuadquizaminosWeb.ContestRules) %> >Contest rules</a></li>
          <li><a href= <%= Routes.live_path(QuadquizaminosWeb.Endpoint, QuadquizaminosWeb.TermsOfServiceLive) %> > Terms Of Service </a></li>
